+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## v2.0.0-beta.2 - 2024-02-27
+
+- Enable federation of a sub catalog path.
+- Enable filtering of federated resource catalogs.
+
 ## v2.0.0-beta.1 - 2023-04-24
 
 Follow Nexus changes.